third_party/WebKit/LayoutTests/fast/dom/shadow/content-pseudo-element-with-host-pseudo-class-2.html

<!doctype html>
<html>
<head>
<script src="resources/shadow-dom.js"></script>
<script src="../../../resources/js-test.js"></script>
</head>
<body class="bodyClass">
  <div id="sandbox"></div>
  <pre id="console"></pre>
</body>
<script>
description("Test for crbug.com/274063: cannot style ::content with a rule that includes :host.");

var sandbox = document.getElementById('sandbox');

sandbox.appendChild(
    createDOM('div', {'id': 'host1'},
        createShadowRoot(
            createDOM('style', {},
                document.createTextNode(':host > div > ::content > * { color: green; }')),
            createDOM('div', {'id': 'host2'},
                createShadowRoot(
                    createDOM('style', {},
                        document.createTextNode(':host > ::content > * { color: red; }')),
                    createDOM('content', {'id': 'content2'})),
                createDOM('content', {'id': 'content1'}))),
        createDOM('div', {'class': 'content'},
            document.createTextNode('content'))));

shouldBe('window.getComputedStyle(document.querySelector("#host1 > .content")).color', '"rgb(0, 128, 0)"');
</script>
</html>
