third_party/WebKit/LayoutTests/fast/dom/shadow/content-pseudo-element-with-host-pseudo-class-2-expected.txt

content
Test for crbug.com/274063: cannot style ::content with a rule that includes :host.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.getComputedStyle(document.querySelector("#host1 > .content")).color is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE
